make a front end but isnot full

+++ src/app/components/vehicle-list/vehicle-list.component.ts
@@ -1,3 +1,4 @@
+// vehicle-list.component.ts
 import { Component, OnInit } from '@angular/core';
 import { VehicleService } from '../../services/vehicle.service';
 import { Vehicle } from '../../models/vehicle.model';
@@ -10,29 +11,53 @@ import { Router } from '@angular/router';
 })
 export class VehicleListComponent implements OnInit {
   vehicles: Vehicle[] = [];
+  page = 1;
+  pageSize = 10;
+  totalItems = 0;
+  searchTerm = '';
+  searchYear = '';
 
   constructor(
     private vehicleService: VehicleService,
-    private router: Router
+    private router: Router,
   ) {}
 
   ngOnInit(): void {
     this.loadVehicles();
   }
 
-  loadVehicles() {
-    this.vehicleService.getVehicles().subscribe((vehicles) => {
-      this.vehicles = vehicles;
+  loadVehicles(): void {
+    this.vehicleService.getVehicles(this.page, this.pageSize, this.searchTerm, this.searchYear).subscribe((response) => {
+      this.vehicles = response.data;
+      this.totalItems = response.totalItems;
     });
   }
 
-  editVehicle(id: number) {
-    this.router.navigate(['/vehicle-form', id]);
+  onSearch(): void {
+    this.page = 1;
+    this.loadVehicles();
   }
 
-  deleteVehicle(id: number) {
-    this.vehicleService.deleteVehicle(id).subscribe(() => {
-      this.loadVehicles(); // Recargar la lista después de eliminar
-    });
+  viewVehicle(id: number): void {
+    this.router.navigate(['/vehicles', id]);
+  }
+
+  // Método para editar un vehículo
+  editVehicle(id: number): void {
+    this.router.navigate(['/vehicles', id, 'edit']);
+  }
+
+  // Método para eliminar un vehículo
+  deleteVehicle(id: number): void {
+    if (confirm('¿Estás seguro de eliminar este vehículo?')) {
+      this.vehicleService.deleteVehicle(id).subscribe(() => {
+        this.loadVehicles(); // Recargar la lista después de eliminar
+      });
+    }
+  }
+
+  // Propiedad calculada para el total de páginas
+  get totalPages(): number {
+    return Math.ceil(this.totalItems / this.pageSize);
   }
 }
